fake-news-pipeline/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
joblib>=1.3.0
pyyaml>=6.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
httpx>=0.25.0
pytest>=7.4.0
flake8>=6.0.0
